requirements.txt
python-dotenv==1.2.1
discord.py==2.6.4
beautifulsoup4==4.14.3
requests==2.32.5
prettytable==3.17.0
yt-dlp==2025.11.12
PyNaCl==1.6.1